advanced tx view added, bugs fixed
